+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.4.2
+- Added lusca for security, enabling CSRF by default
+- Added grunt-nsp-package for module vulnerability scanning during `package` task
+
 ## 0.4.1
 - Fixed an edge case in `hypermedia()` such that the `URI` is not lifted into the `links` Array via automagic, updated test
 - Updated `hypermedia()` to add a `collection` link for an `Entity` representation _[RFC6573]_, updated test